+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.0
+
+- `PoolExecutor` implements leastcon selection
+- `PoolExecutor` is extensible to use implementations of `Executor` other than `IsolateExecutor`
+- requires flutter 3.0
+
 ## 0.2.1
 
 - support reentrancy